% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rmst2.R
\name{print.rmst2}
\alias{print.rmst2}
\title{print.rmst2}
\usage{
\method{print}{rmst2}(x, digits = 3, ...)
}
\arguments{
\item{x}{Object to be printed.}

\item{digits}{Integer indicating the number of decimal places.}

\item{...}{Further arguments ignored in this function.}
}
\value{
returns summary output for class 'rmst2'
}
\description{
S3 method for class 'rmst2'
}
